+++ test/html5/cypress.config.js
@@ -1,0 +1,10 @@
+const { defineConfig } = require("cypress");
+const getCompareSnapshotsPlugin = require('cypress-image-diff-js/dist/plugin');
+
+module.exports = defineConfig({
+  e2e: {
+    setupNodeEvents(on, config) {
+      getCompareSnapshotsPlugin(on, config);
+    },
+  },
+});